CONN-10486 SnowflakeSinkServiceV2IT refactor
